adding test units for exercises from 2 to 7

+++ EulerProject/Algorithms/2.cs
@@ -9,11 +9,11 @@ namespace EulerProject.Algorithms
     public class _2
     {
         //By considering the terms in the Fibonacci sequence whose values do not exceed n, find the sum of the even-valued terms.
-        public static long fiboEvenSum(int limit)
+        public static int fiboEvenSum(int limit)
         {
             if (limit == 2) return 2;
             List<int> list = new List<int>() { 1, 2 };
-            long total = limit > 2 ? 2 : 0;
+            int total = limit > 2 ? 2 : 0;
             while (list.Last() <= limit)
             {
                 int sum = list.Last() + list[list.Count - 2];
